Mobx > Initial - documents folder refactoring
